Move frontend tests to src folder
